Remove most references to libvirt in virt.py
It is still used to specify the network type object, but this is minor.

(Also fix other review comments)
